Patch 2.6.2 for "modern" compilers and libraries
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/generator/branches/R-2_6_3@6515 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/Numerical/Simpson1D.cxx
@@ -64,7 +64,7 @@ double Simpson1D::Integrate(GSFunc & gsfunc) const
   // integral value converges to the real one within the required accuracy
   for(unsigned int iter=0; iter<fIMaxConv; iter++) {
 
-    np = (unsigned int) TMath::Power(2,n) + 1;
+    np = (unsigned int) TMath::Power(2, (int)n) + 1;
     n++;
 
     fmap.IncreaseGridDensity(np);
